Support for module-map only packages
Module maps require absolute paths, so the referenced libraries must be installed on the system.

The module map can describe multiple modules in the same “package”, but a module is a single dynamic library and all the headers that define the symbols in that library, and strictly no more.

Headers must not be included multiple times, so don't specify foo.h if is included by bar.h, just specify bar.h

We have plans to deal with the absolute path issue, since it restricts how flexible this system can be.

We have no tests for this feature as yet, because in order to test it we need to build some C libraries, and the set up for such tests is time consuming.