If we abort during get allow resumes
Note that this isn’t really sufficient, we need to at least fetch and hard reset too.

However during local-package-dev this step is a common occurrence where the clone succeeds but the repo is not tagged. So this is a first step towards correcting that, will have subsequent commits that handle the remaining edge cases.