Ensure we get all tags during initial clone
If a tag is not on the master branch then its ref will not be visible after the initial clone.

A repeat of swift-build after such a failure would execute a `git fetch` on the clone thus fetching that tag.

This diff makes the behavior the same for both paths.

Fixes rdar://problem/23666291