[swiftpm][tests] Rely on XCTest for xplat shim
The package manager tests perform some interesting acrobatics in order
to support unit tests on all platforms while also enforcing type safety.
Change this to rely on swift-corelibs-xctest as a shim layer, as that
library intends. This is the first step to using `XCTMain()` from within
the package manager tests.

swift-corelibs-xctest provides an API for developers to write unit
tests using XCTestCase. Since Swift XCTest cannot automatically detect test
methods like Obj-C XCTest can, Swift XCTest defines XCTestCase as
conforming to XCTestCaseProvider, which defines the `allTests` public
variable. The shim layer exists in swift-corelibs-xctest, and this is
where is should be.

The package manager currently explicitly marks all of its test cases as
conforming to XCTestCaseProvider. XCTestCaseProvider isn't available on
all platforms, so the package manager tests use macros to conditionally
define it. The logic for this conditional definition is convoluted:

- All test cases conform to XCTestCaseProvider, regardless of platform.
- XCTestCaseProvider.swift defines XCTestCaseProvider as an empty
  protocol if the current platform is OS X (which uses Obj-C XCTest and thus
  doesn't define XCTestCaseProvider). This file is not included in any
  of the test targets, so it is not included when building from the command
  line on OS X or Linux. It is included in the Xcode project, thus
  allowing tests to build in Xcode.
- Each target's main.swift defines XCTestCaseProvider as a protocol
  that defines the `allTests` public variable, but only if the platform
  is not Linux (on Linux, XCTestCaseProvider is already defined). These
  files are used when building from the command-line, but are not
  included in the Xcode project (thus the Xcode project does not have
  duplicate definitions of XCTestCaseProvider).

The benefit to doing all this is to ensure that all package manager
tests conform to the XCTestCaseProvider protocol, and so developers
are informed at compile-time whether their tests will work on Linux.
This is great, but not worth the complexity above.