Better error description for typos in local repository paths
If the path to a local package repository contains a typo, the error message is not particularly helpful. Consider this `Package.swift`:

```
let package = Package(
  dependencies: [
    .Package(url: "../CJPG.swift", majorVersion: 1)
  ]
)
```
If the actual repo is called `CJPEG.swift` then the following error will be generated by `swift build`:

`swift build: realpath error: No such file or directory (2)`

This may confuse users and is especially difficult if there're a lot of dependencies. Printing the offending path in the error case should clear things up.