Document "the output is a binary" more clearly.
The word "package manager" is overloaded and often refers to things that
try and *do something* with the built software (usually install it,
e.g. `apt`, `homebrew`, `pod` `pip`, etc.)

swiftpm's role begins and ends with building a binary, which may be
unexpected from the project name alone.  This design decision is
important enough to document in the README, so that there are no users
poking around in `/usr/local/bin` wondering where their package got
installed.