Add SWIFT_PACKAGE define
swiftpm should really define a preprocessor macro directive so that the
source may configure itself appropriately for the package manager.

More specifically, I wanted to do

    #if SWIFT_PACKAGE
    import Foo
    #endif

Because the (formerly-built-with-xcodebuild) library I'm packaging has a
bridging header that imports Foo in some other way, but bridging headers
aren't supported in SwiftPM so I'm importing that at the top of various
files.

Add documentation related to this feature

Add tests related to this feature