[bootstrap] Refactor to be compatible with Python 2 or 3
The desire was expressed in MR #59 that the bootstrap script should be
refacotred to be compatible with both Python 2 and 3 rather than be
hardcoded to one only. This patch refactors the bootstrap script
to do just that.

I was able to test this on OS X (`Python 2.7.11`) and on Arch Linux
(`Python 3.5.1`).