[Python] Add flake8 config
In https://github.com/apple/swift/pull/1153, @practicalswift added a
flake8 config to the Swift project, in order to lint the Python scripts
used by that project. This does the same for the package manager.

This includes all the linting rules active on the main Swift repository.
It also includes tweaks to `Utilities/bootstrap`, which bring this
project in full compliance with those linting rules.