Add -Xswiftc for passing args to swiftc
Like our -Xcc and -Xlinker flags, -Xswiftc is only allowed on the command line and not part of Package.swift in an effort to prevent CMake/Autotools like developer hells.

Strictly we add these flags so developers can work around issues, but they also should report a bug so that we can provide a proper solution for their needs.