Less global state (-Resources.swift)
This global state was making everything less testable, in the end I realized mostly it was there for ManifestParser, so now the paths to that are injected.

A bonus is that Get no longer depends on ManifestParser which is a more restricted dependency diagram and makes the module more testable too.

I also removed the code that figures out the Darwin path to the executable using dlsym. Perhaps this is foolish, but `Process.arguments.first!` seems as good and doesn’t have an initialization order problem, and since XCTest exposes no way to statically initialize global state, I don’t want to deal with it.

——————

Though I am not so happy with the injected manifestParser() function in Get. It feels dirty to pass it around to every object, but thus is the pattern.