[bootstrap] Preemptively add argument for Foundation
https://github.com/apple/swift-corelibs-xctest/pull/43 will
add a dependency between swift-corelibs-xctest and
swift-corelibs-foundation. This will necessitate the path to
a Foundation build be passed to the SwiftPM bootstrap script.

In order to prevent CI from breaking, we'll modify the Swift
build script to pass the correct parameters *before* those
parameters are actually used. To do that, this commit ensures
the bootstrap script accepts the new parameter: "--foundation".