getcwd() no longer throws
This is an example of too-edge-case a runtime error to throw IMO.

Instead there is a safe way to manage the process working directory so that the single error that is possible, ie. there is no working directory, cannot happen.

So we document that you should follow that methodology and instead fatalError in the event that the error does occur.

The consequences here are many situations where try was added to simple functions, and those trys have been removed.