Can build with toolchain other than swift-build’s
This is an essential part of the puzzle that is:

* swift-build requires Swift 3 to build and consume Package.swift
* Many packages exist that require Swift 2.2

The workaround here is to set `SWIFT_EXEC`, swift-build will use its toolchain for the parts that are intrinsically coupled to the toolchain, but composes a build-YAML that uses the toolchain with which `SWIFT_EXEC` is a part.

Since SWIFT_EXEC is an expected feature, we are adapting it here, but it may make more sense long-term to have a dedicated `--toolchain` flag.