[SR-1540] Allow bootstrap script to take a human-readable version string
This adds a new --version-string option to the
bootstrap script.  When set, this causes the
bootstrap script to emit a header containing a
function that returns the string literal, and a
module map so Swift can find it.  main.swift is
also modified to conditionally invoke this func-
tion to get the string to print out.