[SE-0085] Add a `package` subcommand to handle package-oriented operations.
This moves the package-management commands from the `build` subcommand to
`package`, but leaves building (and, currently, cleaning) on the `build`
subcommand.  The `test` subcommand is likewise unmodified.

I had to temporarily move down the enums for InitMode and ShowDependen-
ciesMode, so that both package and build can share them, but this will
be consolidated again when build and test are reimplemented in terms of
package as well.  The build and test functionality is not yet moved in
order to minimize changes while still ending up with the desired func-
tionality.