Revert "[Tests] Remove an unnecessary sleep()."
There was a non-deterministic failure in this test following this
removal. Reverting for now while I investigate to see how it could be
that llbuild is not detecting the change in this situation (the change
should have altered the file size, which should make it resilient to
timing related issues in the filesystem, but that apparently is not the
case).

This reverts commit 15d4e6c14cf08d81e644249880a1bb5f05fe06cf.