Add the start of AbsolutePath and RelativePath types.
This will provide a more type safe and expressive path API than plain
strings.  Many more details are in the documentation of the types.

More extensive correctness and performance tests, as well as adoption,
to come.  There is still a fair amount of performance investigation and
tuning to do -- that can be addressed piecemeal, but the direction of
the API (in particular the distinction between absolute and relative
paths) is the most important part at this point.