Fix bootstrap script to add SWIFT_EXEC to xcproj in a more robust way.
When code coverage was enabled earlier today, this broke the bootstrap
script's (admittedly fairly hacky) way of injecting a scheme assignment
of the environment variable SWIFT_EXEC in a generated Xcode project.

This caused all functional tests in generated projects to fail in quite
mysterious ways.

The new way of adding SWIFT_EXEC is more robust, and now the check for
whether or not adding it was successful works correctly, so we should
know sooner if it breaks again.