First round of String -> Path transition.
Among other things this adds temporary shims to FSProxy until all the
clients have been converted to use path types.

Note that during the conversion there are a lot of .asString calls that
are needed for POSIX functions and FileManager functions, but as we put
all of these on FSProxy (which takes Paths) a lot of these should end up
disappearing again.

So during the transition, the current path-based code looks a bit more
complicated than it will eventually look.  The end goal is for the path
operations to be type safe, elegant, and intuitive.