[Bootstrap] Add C language support in bootstrap script
This adds a very basic support for bootstrapping C targets in swiftpm. It doesn't have all the features yet for eg: modulemap generation and C executables.
A shared library is created for C targets and placed inside the lib folder. Include path to modulemap location is added to the swift targets which depends on C targets.