Remove `ExpressibleByStringLiteral` conformance for AbsolutePath. This was used to make transition simpler, but can cause confusing runtime errors (since there are implicit requirements about whether or not the string starts with a `/` depending on the context).
There are a fair number of places in which explicit `AbsolutePath(s)`
construction has to be added.  However (and this is important):  the
vast majority of the cases are in test code, which is in the business
of dreaming up arbitrary test paths based on literal strings anyway.
In the production code, there is far less of that needed (almost none).
It's alright if the unit tests are a bit word, it's the business logic
that we want to look nice and clean.  For unit tests was always add
more convenience methods to make things cleaner without polluting the
actual underlying API.