[Build] Consider Swift on Clang Modules as build order dependency
Currently swiftpm fatalError() when some swift package is dependency
on a package containing c languages or mixed languages. This shouldn't
error out. Swift will not be importable to c modules but there could be
swift modules in that package which want to use some external deps.
Considering it as a build order dependency should be ok.