Remove testDependencies from PackageDescription
This feature was supposed to support dependencies only for the root
package but at some point it stopped working however the API still
remained open. This patch removes the public API. This is a valid and
desired feature which is supposed to come back after it goes through
proper review on swift evolution.