[PackageDescription] Add a JSON type
A JSON type similar to Basic.JSON but a very minimal version.
It will be useful and nice to use this instead of direct string writing
for serializing manifest files. At later point if we ever link
PackageDescription with Basic. it can be easily removed and Basic.JSON
should be easily plugged in.