Reorder swiftc args to allow user override
Currently the build target arguments are hard coded in the swiftpm
source. Until they are broken out into some configuration method, it
would be useful for users to be able to override these arguments for
their use case. This change reorders the passed arguments to pass along
the users arguments after the default arguments.