Enable the bootstrap script to extract product definitions from SwiftPM package manifest. This is preparatory work for an upcoming improvement, but can stand on its own.
This also cleans up the contents of SwiftPM's package manifest itself.

Note:  The bootstrap script is one big hack, and we don't have unit tests for it.  That's why this diff has no tests.  This is not ideal, but is not horrible, since the bootstrap script is used under controlled circumstances and does not become part of the product.  The only package on which it is ever used is the SwiftPM package itself.