[SourceControl] Add basic in-memory git implementation
This implements a basic in-memory implementation of git. Currently,
It will be useful in workspace tests which uses real filesystem for
git operation. Using in memory implementation makes the tests around
22.5 times faster which means we can test much more number of things
without worrying about the increase in test time. We anyway add
functional tests in all our tools so we will not be losing any test
coverage.