[Build] Manually link ClangModules in Swift products
This will get rid of current autolinking for linking Clang modules in
Swift modules. This will help when the product proposal is implemented, i.e.
when we statically link ClangModules we won't need to create .a files.

This also means that all users using a custom modulemap will need to
remove the link statement from their modulemap.

- <rdar://problem/30016942> Link C language libraries via Build instead
  of using autolinking via modulemaps