[Xcodeproj] Set defines module to NO for ClangModule
We were generating a broken framework for Clang modules because:
1. We don't have a copy header phase.
2. The module map doesn't contain "framework" keyword in it.

This was not breaking the builds but it broke autocomplete and
iterface generation when using the generated Xcode project.

The workaround includes setting defines module to NO and use explicit
flag -fmodule-map-file to pass the path to modulemap.

This PR also includes a workaround for <rdar://problem/21912068>
SourceKit cannot handle relative include paths (working directory)

- <rdar://problem/29825757> SR-3121: Xcode claims it could not build
  Swift-packaged Objective-C module even though it did