[PackageDescription] Restore the old product API
This API was removed when implementing the product proposal but is being
restored to keep compatibility with current Swift 3 packages. The new
API will be introduced in the PackageDescription4 module.

This PR removes some unit tests which are not worth re-adding however
we still have functional tests around the 'working' product APIs.