Add PackageDescription4 module
This will be the new PackageDescription API in Swift 4. This is named
PackageDescription4 because we already have a PackageDescription.
However this will be installed as PackageDescription in Toolchain so
clients will actually use `import PackageDescription` when using this
version of manifest. This will need to be recompiled and installed using
the bootstrap script as PackageDescription.