SR-4083 Generated projects should have an aggregate target for every product that doesn't have any other target
This adds to the Xcode project generation logic so that, after adding all the targets for the modules in a package and its dependencies, it also adds an aggregate target for each of the products that doesn’t already have a target with the same name.  This makes it possible to build all the modules that go into a defined product in a single build in Xcode.