WIP: A start at a resillient PinStore that holds onto error rather than thowing in init()
Currently, many of the critical objects involved in the Workspace throw errors in their initializers.  Since they are non-optional in the higher-level objects, those initializers have to throw as well.  The result is that, if there is any error anywhere in the graph, it isn’t possible to instantiate a Workspace.

This is a serious problem for libSwiftPM.  The approach, of which this WIP branch is a start, is to make the initializers not throw but instead record the errors.  Objects with errors can still be queried, but any attempt to perform operations on them (especially operations that could end up overwriting files etc) will throw an error that informs that client that the operation can’t be performed because of a prior error, and that wraps the original error.

We will definitely need to discuss the form of the errors (i.e. we should have a type for diagnostics, and should be able to handle things like note/warning/error distinctions etc.  We also need to discuss what a protocol should look like that encapsulates some of this functionality and APIs.

PinsStore is the first type to be converted since it is a leaf type.  The others, up to and including Workspace itself, need to be similarly modified.  We can do that by extending this PR or through new PRs.