[BuildPlan] Strip invalid flags from link command
This strips the -wmo flag from link command if it was manually passed
during swift build invocation. It causes compilation failure right now.
We might not want to keep this long term once compiler strips the flag
itself so this is workaround for now. Having wmo in debug mode is a
valid use-case which we plan to properly support using build settings in
future.

-- <rdar://problem/30985305>