[PackageGraph] Add incomplete mode to resolver
This mode lets the resolver run without fetching the containers which
it doesn't already have. This is useful when we want to run the resolver
with different sets of constraints without going off cloning a lot of
new repositories which might be unnecessary.

This, with delta debugging algo, will eventually let us diagnose the
unresolvable package graphs.