Avoid using the system module cache directory for SwiftPM's test
This will override the module cache directory with the path in env variable for
SwiftPM's tests. We do this because each test fixure creating their own module
cache will take a lot of time. This particular change is being done because
there are compatibility issues in using the shared module cache location on the
SwiftCI across compiler versions.

-- https://bugs.swift.org/browse/SR-4474