Migrate Workspace.clean to the diagnostics system
Made Workspace.clean non-throwing and accept a `DiagnosticsEngine`.

While doing so, I found it quite useful to define a `wrap` method on `DiagnosticsEngine` that allows us to wrap a throwing function into a Optional returning function, feeding the engine on error. It’s a more generally useful version of `LoadableResult.load(diagnostics:)`, so took the liberty to remove that function.