Introduce Package Identity
Package identity will be used to uniquely identify packages during
dependency resolution.

By default, the package identity is the basename of the package
repository URL. In future, the package references in the manifest file
will be allowed to override the name that should be used instead of the
basename. The package identity should always match (case insenstive for
now) the name of the package. This will be enforced via error checking
in future, i.e., once we have the API for overriding.

The commit will make dependency resolver to use the package identity
instead of package URL. The package graph is still constructed using the
package URL but that will be updated as well in upcoming commits.