[PackageGraph] Add support for "local" packages in resolver
A local package is a package present on the file system which may or may
not be under a git repository. These packages should be used as-is and
should not be cloned or managed by SwiftPM. Example of such packages
are: root, edited, local dependencies and multi-package packages.

So far, we only cared about edited packages during the resolution and
used unversioned requirement to handle them. This commit extends that
logic but instead of pre-loading the requirements from an unversioned
package, the container API now has a method to fetch dependencies of
unversioned package.