[BuildPlan] Always add -fblocks when compiling C targets
<rdar://problem/28109524> [SR-6178]: Pass -fblocks to clang for C language targets

Currently, blocks in C libraries will only work if the C target is being
imported by a Swift target and that Swift target imports Dispatch. This
isn't very optimal but its better than passing -Xcc -fblocks on the
command line. This will not cause any issues if the C target is not
using blocks. We can track always linking against blocks runtime as an
enhancement.