Implement llbuild manifest caching
This uses the new directory structure feature of llbuild to compute if
manifest generation is required. We use a command to write a token file
inside the build directory and regenerate the manifest file if the token
file is written. This saves a lot of time in builds where directory
structure hasn't changed. This feature is off by default and can be
enabled by using the flag --enable-build-manifest-caching.

<rdar://problem/36365296> Implement build manifest caching using llbuild's directory structure feature