[PackageGraph] Cache resolved subtree assignment set sequences
This avoid a lot of unnecessary work done by the resolver when there are
shared dependencies in a package graph. The resolver resolves the
subtree recursively everytime a dependency is encountered, so this
provides very very good performance in shared dependencies cases.

This patch was originally written by Daniel Dunbar.