R-37354691
Make sure that OTHER_CFLAGS starts out with `$(inherited)`, as OTHER_LDFLAGS and OTHER_SWIFT_FLAGS do.  This is appropriate even if there is no other addition to OTHER_CFLAGS; in that case it is a no-op but serves as an indication that if anything else is added by hand, `$(inherited)` needs to be included to avoid masking underlying values.

(cherry picked from commit b5542af9aa7c6a185bcba110a7166f2d1f3410a9)