[PackageGraph] Fix hashing for PackageContainerConstraintSet
The ordering of its elements is not an essential part of a Dictionary;
two dictionaries with the same elements compare equal even if they're in
a different order. We must also produce hash values independent of element
ordering, or we will run afool of Hashable requirements.

Re-enable the test that was failing nondeterministically in previous commits.

rdar://problem/38647101