SR-6277 "WIP" Mute non-json output from dump-package
Submitting PR for feedback on the solution, in advance of writing unit tests.

This PR mutes tool progress output by default.

It also introduces a new dump-package option: ` --include-tool-output` to override this default and show dependency resolution and fetch progress as before.

Diagnostics that contain an error are still printed to the console.

I am seeking feedback on this solution as I an new to this codebase and unfamiliar with it's broader structure, and expectations regarding tool putput handling.