[Build] Pass object files via a file to Swift compiler
<rdar://problem/38803919> Unable to compile the code via terminal using swift test

This is required for packages with large number of dependencies which
may cause the command line invocation to surpass the arg max limit. The
previous commit was reverted because the response file is supposed to be
space separated and not new line separated. I also updated our
in-repository test that checks spaces to actually build a product.