Emit a script allow self-hosting SwiftPM after bootstrap
<rdar://problem/42350010> Emit a script allow self-hosting SwiftPM after bootstrap

SwiftPM allowed using the built/inferior swift-{build, package, run, test} to
iterate on the package manager codebase. This is going to break now because we
adopted llbuild as a link time dependency. To resolve this, the bootstrap
script will emit a script that will call the swiftpm tool with the additional
flags that are required to allow self-hosting. This can be removed once we have
actual support for build settings.