[Commands] Remove build manifest caching
<rdar://problem/42583418> Remove build manifest caching

The build manifest caching provides a preformance win but creating build
plan is now faster (than build manifest caching) because of package
manifest caching. The difference is about 100ms and we can bring the
time down by optimizing the tools lookup. However, this feature is not
critical anymore because the performance is decent with package manifest
caching. We can add it back in the future when required.
￼