[PackageLoading] Print warnings when loading manifests
We were supressing the warnings while loading the manifests because we
considered any type of output from the compiler as a sign of error. This
changes the error state to depend on the exit code. The warnings will be
printed when loading a resolved package graph and not during dependency
resolution. This makes sense because the manifest we look at during
dependency resolution might not be used in the final resolved graph.

<rdar://problem/41180566> SwiftPM should not suppress compiler warnings in the manifest file