Link back to full docs.
Search results and my desire to use the latest version made me think that I could find everything I needed to use in the 4.2 documentation.

Come to find out most it and the specs are in PackageDescriptionV4.md with 4.2 documentation being a supplement. 

Thinking a link near the top might help call this out. Or some copy mentioning that this is a modification to the full set of docs.