[PackageGraph] Use OrderedDictionary to store assignments
The assignments were stored in a regular dictionary which caused
non-deterministic crashes when a package graph contained dependencies
that were required via a revison and version in different subtrees.
Since the order in which the constrained appear matters, the assignment
also needs to be stored in order.

Cherry-picked from 1c482a5d

<rdar://problem/43699736>