[Workspace] Resolve correctly with outdated Package.resolved file
This fixes the issue when we would end up using a dependency URL from an
outdated Package.resolved file for a transitive dependency. Since the
resolver clones the first package identity it encounters, we would not
clone the right thing and then would end up in a broken state in later
stages.

The patch will basically re-run the resolver once we have enough
information to correctly determine the correct path of the trasitive
dependency.

<rdar://problem/44948545>