Prefix public stdlib extensions with spm_
Extensions are very easily leaked into the clients. Since Swift doesn't
have the concept of private imports at this point, it is not safe to
have public stdlib extensions as it can easily cause collisions in the
downstream clients even though they might not directly depend on
SwiftPM. This patch prefixes most of the public stdlib extensions in the
SwiftPM codebase. All future additions should also have the prefix.

<rdar://problem/45200636> SwiftPM should not be exporting Swift stdlib additions