[Commands] Add an option to force resolution to Package.resolved
This adds a new flag `--force-resolved-versions` which will force
resolve to versions recorded in the Package.resolved file.

The entries in Package.resolved can be considered as "soft" pins.  Since
the requirements in root manifest and Package.resolved can change
anytime, most SwiftPM operations (that load the package graph) first
check if the current state of things is still correct. However, it is
sometimes useful to force resolve to the dependencies recorded in the
resolved file by entirely bypassing the dependency resolution.

<rdar://problem/45290071> Add option to not re-resolve, but use Package.resolved explicitly