[Build] Add support for indexing-while-building
This add supports for indexing while building which will be useful for
LSP users. The indexing-while-building feature will be enabled by
default when building in debug mode but can be disabled by passing the
flag --disable-index-store. The feature is never enabled in release.

This feature is supported by Swift compiler on all platforms but clang
has some nuances. We can expect this feature to be available on Apple's
clang but not OSS or other distributions of clang. So, the feature is
selectively enabled for clang targets.

<rdar://problem/45689037> Add support for indexing-while-building