[Basic] Add ArraySlice<UInt8> specialization for OutputStreams
LocalFileOutputStream copies collections into [UInt8] before calling
fwrite in writeImpl. This overhead can be avoided when an ArraySlice is
available. Further, the write() method is always buffered, which
potentially incurs another copy overhead. This change adds an
ArraySlice<UInt8> specialization for OutputStreams and adds an
'unbuffered' option for the class. The combination of these two changes
yielded a 10x performance improvement in a tool utilizing these.

rdar://problem/45592874