[Utility] Add a cross-platform filesystem watching utility
This adds a cross-plaform filesystem watching utility based on
FSEventStream (macOS) and inotify (Linux). The eventual goal is to add
APIs in libSwiftPM so long living clients like IDE can be notified of
changes in the manifest or the file structure of the package.

The inotify wrapper needs some work but that can happen in follow on
commits.

<rdar://problem/45795219> Add APIs for watching the filesystem for changes