[PackageLoading] Invalidate manifest cache when env changes
Many packages use env variables to perform conditionals in the manifest
file. This basically means that we need to invalidate the manifest cache
if the env changes. Ideally, SwiftPM should provide an API in
PackageDescription that will allow us to keep track of the exact
variables used by the manifest but this should be good enough for now.

<rdar://problem/45931785> SwiftPM should invalidate manifest cache when the process env changes