[bootstrap] Remove PackageDescription linking hack
SwiftPM's bootstrap script derives the linking command by calling
swiftc with -###. I think this was needed in the past because we
couldn't forward the -Xlinker args to the linker on linux. This no
longer seems to be true since Swift is using clang++ on linux as the
linker. The -Xlinker args are correctly forwarded to clang++ so we can
stop using this hack. This hack surfaced an issue when trying to land
the Codable in PackageDescription PR as we were dropping the swiftrt.o
file during linking.

<rdar://problem/45976814> SwiftPM shouldn't need to derive the linking command when producing its runtimes on linux