[PackageDescription] Add target-specific build settings API
This adds the build settings API that is currently being pitched on
Swift evolution. The intent is to add the required infrastruture for
build settings even if we don't end up with an accepted proposal before
the next Swift release. The APIs currently have an underscore to
indicate that they're not officially supported and will be removed if we
don't go through the evolution process.