Improve OS version number validation in SupportedPlatforms
The version string of an Apple platform is a dot-separated series of
2 or 3 positive integers. This improves the validation to ensure that
the version components are non-negative, and that the version number is
within the range of publicly-deployable versions that actually existed
for that platform (this is different from the minimum *supported*
version).

For example, the minimum versions in this validations are the minimum
versions that any API can be documented as being available on:
- iOS 2.0+
- macOS 10.0+
- tvOS 9.0+
- watchOS 2.0+

This also deduplicates the code.