[bootstrap] Allow modifying the stdlib RPATH using a build where the stdlib is symlinked into the toolchain
This is a simple change to use 'normpath', for removing '..' from the stdlib path, instead of 'realpath',
which allows installing swiftpm using a setup where the Swift stdlib is symlinked in.
Without this change RPATH modification fails because the real path string is not found in the binary
(the absolute stdlib RPATH in the binary is relative to the swift binary).

This should have no effect in existing setups.