[PackageLoading] Improve error message when a custom target path does not exist (#1944)
* [PackageLoading] Improve error message when a custom target path does not exist

This fix SR-9558.

The issue is that even when a custom target path is specified in manifest,
it still shows an error message asking to specify a path. The fix is to check
if a path is already specified but does not exist, then it indicates the path
dose not exist. A test is added for the fix.

* Improve style and error message.

* Futher improve style and error message.