[PackageGraph] Improve handling of branch-based dependencies
This improves handling of branch-based inside the resolver which leads
to a deterministic handling of the revision that is picked as part of
the resolution process. This also fixes the bug that branch-based
dependencies were being always resolved from HEAD (if there was no build
cache) when a resolved file was present.

<rdar://problem/45289738> Dependency resolution doesn't respect the resolved file for branch-based dependencies when starting resolution process from scratch