[Commands] Add preliminary support for swift-format
This adds just a simple `swift package format` command to format all the
sources in your root package. The swift-format binary will be looked up
using the env variable SWIFT_FORMAT_EXEC.

This patch is just for testing the proposed swift-format tool. We will
need to figure out the specifics of the integration if/when the proposal
is accepted.