[PackageGraph] Use basename as identity of root packages
Using root package's name as the identity doesn't really work properly
because there is no way to override the basename in the urls for
package dependencies. This does require that root package's basename
is same as the dependency url's basename but that is better than having
no solution for this.

<rdar://problem/48520076>