[SPMLLBuild] Handle fatal errors from the build engine
We were not handling fatal errors from the build engine that can happen
in rare situations (such as the db being deleted while a build is in
progress). In those cases, llbuild returns an empty value for the
overall result and reports errors via the delegate. This implements the
delegate and throws the errors when this happens.

<rdar://problem/49337286> Handle fatal errors from the build engine