Add new `manifest` option to `swift package init` (#2106) (#2108)
This will just generate a manifest in the current directory. The
contents will be the same as for initializing a library package.

That means the contents will likely not match the filesystem layout,
but I think that is OK, this will still make it easier to add a
manifest to an existing project. We can improve it in the future and
make the generated content based on what is actually there on the
filesystem.

rdar://problem/50347943