add -DCROSS_COMPILATION_TARGET_LINUX to compiler invocations
Although it's not recommended, many packages use `#if os(...)`
conditionals in their `Package.swift` which assumes that the OS running
`Package.swift` is the same as the one running the resulting binaries.
For cross-compilation that's not true. By passing
`-DCROSS_COMPILATION_TARGET_LINUX` to all compiler invocations, package
authors will be able to change their

    #if os(Linux)

to

    #if CROSS_COMPILATION_TARGET_LINUX || os(Linux)

and we have at least a workaround...